Print out git output on merge errors
A comment in the code suggests that these errors were already being
printed out, but that seems not to be the case.  I believe it used to
be true when the comment was added but the refactor in aosp/1319045
accidentally broke it when it moved from a subprocess.run call that
prints the output to stdout to subprocess.check_output, which returns
it instead.  So let's print the output ourselves.

Test: Run on something that fails and see the output.
Change-Id: I9413a8237ef4514cfaf25bf3bc6b9e212eac62d1