AIDL_FROZEN_REL/AIDL_FROZEN_OWNERS: sim REL
Add way to simulate the branch being in a release state, at least for
AIDL finalization. This is in preparation to get a test in the tree that
ensures that the single command to freeze AIDL interfaces can be run
without breaking anything in the tree.

AIDL_FROZEN_REL - simulate REL. During release, we consider all
interfaces which have no 'owner' field to need to be frozen. So, if they
are used, it is considered an error (certain interfaces which are marked
as @VintfStability will be tested at runtime by VTS to be frozen,
whether they have an owner as well).

This means that there is no build check for owned interfaces. There are
a couple of usages for this. For instance, 'owner: "test"' is used for
test interfaces, so that they don't need to be frozen. Similarly, for
interfaces which are in development in AOSP, but they aren't yet ready
for release, commonly we add 'owner: "aosp"', precisely to remove this
build time check (again, if the interface is important to be frozen,
this is tested in compliance tests).

AIDL_FROZEN_OWNERS allows the build checks to be extended to other owner
fields. For instance, if an OEM has marked their interfaces as 'owner:
"some-company"', then in their own release process, they can build with
"AIDL_FROZEN_OWNERS='some-company'" in order to verify these interfaces
are frozen in the build, without having to wait for the compliance
tests.

Bug: 190577319
Test: added to the build + manual
Change-Id: I07b1c7f1ffda725c518ad55eb63fc59c5fe95dd0