ui: Move state source of truth to frontend
Previously the controller worker maintained the
'source of truth' of the state object. In the normal
course of things:
1. The frontend (main thread) 'dispatched' DeferredActions.
2. These actions where postMessage'd the controller worker
3. The controller thread applied the action to the state
   producing patches.
4. The patches were returned the frontend which applied them
   to its local copy of the state synchronising it with controller.


There were two main complications:
a. In a minority of cases actions were also dispatched from
   the controller thread. This necessitated careful logic to
   avoid reentrancy in the controller.
b. This scheme imposed a minimum latency of two IPC hops to any
   action. We worked around this with the introduction of
   frontend_local_state.ts. Frontend local state was still part
   of the state but was also updated synchronously on the frontend
   side with all the complicated merging logic that implied.

This change moves the source of truth of for state to the frontend.

After this CL:
1. The frontend 'dispatches' DeferredActions.
2. These are synchronously applied to the state (producing patches)
3. The patches are forwarded to the controller.
4. The controller applies the patches synchronising its copy.

This solves a. and makes it possible to solve b. by removing
frontend_local_state.ts in a future CL.

Change-Id: I91ab584870768df48d443e76eab2f2d73a7f0b97