tp: handle pid reuse after explict end event
This CL handles a subtle edge case which caused processes created after
reuse to not be updated to the pid map. This had the affect of creating
threads contiunously even when they shouldn't be.

In practice, hitting this bug is extremely rare as it requires process
end tracking to be active but process start tracking to *not* be active.

Alternatively (and much more likely), if a process is accidentally
refferred to after the process is endeded to (e.g. by a rougue ftrace
event), it can cause many threads to build up unnecessarily.

Bug: 192274404
Change-Id: Ie684e656a4238229690814df81b37525560c67b5