Autoupdate of 2 submodules
*system/core updated to 2359145f4c64f2eaccf267dc6a72d401392f56cd
Merge "libsnapshot: Add a source build fingerprint to the update state."
*system/update_engine updated to 4657407a32ba168e9dc49277c1ee20359169af9f
Record the source build fingerprint in SnapshotMergeStats.

This also adds a manual call to WriteState(). If update_engine crashes
or the device reboots, it's possible that certain stats (such as the COW
file stats) will be gone by the time the merge state is recovered.
Forcing a write prevents this.

Note that the write was previously implicit in set_cow_file_size, but is
now explicit before calling InitiateMerge().

Bug: 188909957
Test: manual test
Change-Id: I27cdc72272a9fa4ff5df4968e200fc14f63d0146