Autoupdate of 2 submodules
*hardware/interfaces updated to f4c8ff9bef34092787e97149e38cd1bb052635c8
Add a utility to JSON-format a CSR with build info

We need both the build fingerprint as well as the CSR when uploading
data to the APFE provisioning server. Add a utility function to format
the output as a JSON blob so that it may be easily collected in the
factory in a serialized data format, then later uploaded.

Test: libkeymint_remote_prov_support_test
Test: VtsAidlKeyMintTargetTest
Test: VtsHalRemotelyProvisionedComponentTargetTest
Bug: 191301285
Change-Id: I751c5461876d83251869539f1a395ba13cb5cf84

*system/security updated to 2edc36ce1ce3dc5b610d456573fac2c15531ef73
Merge "Add JSON output to the RKP factory tool"