Multiple fixes for ACTIVITY mode.
Improved searching for classpath jars which might be in an
Apex intermediates directory.  Fail fast with a build suggestion
if not all are found.

Handle d8 changes (single dex file no longer supported as an
output option).

Replace hyphens in package names in manifest as these are not legal.

Add min and target sdkVersion to manifest.xml - prevents warning
popups on recent Androids.

Test: Manual tests
Change-Id: Iaf72d6213f9e3cce17190cc007fe2c2224533f79