abitidy: use elf-symbol-id as symbol-declaration link attribute
abidw creates var-decl and function-decl elements corresponding to
global objects with name, mangled-name and elf-symbol-id attributes.
The latter two are not always present.

In the case of the Linux kernel these are, if present, aways
identical, modulo bugs in older versions of libabigail. For pure C
ABIs, mangled-name will never differ from name.

Semantically, abidw links such declarations and elf-symbol elements as
follows. Each ELF symbol has an implicit id taking one of the three
forms:

- name (if not versioned)
- name@version (if versioned but is not the default version)
- name@@version (if versioned and is the default version)

Exactly those declarations with an elf-symbol-id attribute are linked
to a corresponding symbol. In the case of aliases, this should be the
"main" symbol.

This change switches --prune-unreachable to use elf-symbol-id as the
linking attribute, instead of mangled-name, and keys ELF symbols by
their implicit id. The mangled-name attribute is still used as a
secondary key for --sort.

The overall effect of this change is very small. Certain test XML
files with partial symbol alias suppression no longer lose type
information for the remaining aliases. More generally, this change
will preserve more type information in the case of an unexpectedly
absent versioned symbol or a missing mangled-name declaration
attribute.

Change-Id: I8820a4ee095dac7faee488b58b81c6adbd9fe3bd
Signed-off-by: Giuliano Procida <gprocida@google.com>