Latest in fuchsia flashing changed to look at second latest build
There is a possible issue where the latest uploaded build is not
complete when the flashing script tries to pull the build down,
which results in a partial build being pulled down and flashing
to fail.  By changing it to the second most latest, the build
is guaranteed to be fully uploaded when flashing is called.

Bug: None
Test: Verified build is still pulled down correctly.
Change-Id: I2c69f94eed5ebb9bd5f14dcf7ad5a1ba1426ac82