ui: Remove video viewer
The video viewer feature of the UI has bitrotted over time and recently
added analytics events show zero recorded usages of it since the event
was introduced. Given that it seems sensible to remove the feature.
If we revisit this something more like the Devtools filmstrip seems
like it could be a better UX.

Change-Id: Ia3a44e05310e57af797f65c86fd286975a1224cc