Autoupdate of 3 submodules
*frameworks/base updated to 25b4755974f407c608c76d35ec316404c995f8d3
Merge "Allow usage of SdkLevel in the framework/wifi services"
*frameworks/libs/net updated to f829dd3c64d3f8f5038b9c20886675801519e66f
Support both R and S+ in NetworkFactory

NetworkFactory needs to be backwards compatible as
it is statically linked by the Wifi module, which
needs to run on both R and S.

To achieve this, maintain 2 separate implementations
called NetworkFactoryImpl (S+) and
NetworkFactoryLegacyImpl (R). NetworkFactory itself
becomes a shim that delegates to one of these
implementations depending on the device's SDK version.

CP of ag/14265947

Ignore-AOSP-First: Merge in internal first to fix
                   S-on-R test breakages.
Bug: 183902758
Test: ClientModeImplTest
      OemWifiNetworkFactoryTest
      TelephonyNetworkFactoryTest
      UntrustedWifiNetworkFactoryTest
      WifiNetworkFactoryTest
Change-Id: I6a08f39ad1633dd7c1582c362c97821e24ce6683
Merged-In: I6a08f39ad1633dd7c1582c362c97821e24ce6683

*frameworks/opt/telephony updated to 83e8ee8d79cc5b53aee6cac118ce905a180815ad
Merge "Allow usage of SdkLevel in telephony"