Add proto field to store xor operation in update manifest
When generating OTA, we pre-compute which blocks inside a BSDIFF
can be represented by XOR operation, and store these blocks in
the merge sequence. When installing OTA, delta_performer
looks at the merge sequence, and convert blocks to specified
COW op type. If a block does not show up in the merge sequence,
it is assumed to be COW_REPLACE type.

Test: th
Bug: 177104308

Change-Id: If04ca9cae036f2d3a365dcf92ba2447f7b08005a