Bump CaptivePortalLogin development version to S
Bump CaptivePortalLogin version to 319999900 in the manifest. This is a
development version number, which is replaced by an actual version
number in release builds.

Using 319999900 allows installing CaptivePortalLogin on all S devices
which use 31xxxxxxx prebuilts, and will still allow installing
production T modules (32xxxxxxx) on a development S build. This is
convenient for local testing.

The same convention is also used for other modules including
NetworkStack.

Bug: 193853555
Test: m CaptivePortalLogin, check version code
Change-Id: I9d19fb3c0eb6a073b035e87a88d3891eb631a454