Fixed an output path bug
This CL fixes a bug that was introduced when the build system began
moving the source to the output directory before applying patches or
starting the build process.

One of the stages, copying stdlib source files into the distributable
directory, was left using an old value for the output directory.  This
CL updates the path it uses to the correct one.

Test: build and check output directory structure
Change-Id: If79be1054653666ab0825c857bdcfdad4a4b37ea