Remove the switcher snippet from Dackka
This is due to the switcher snippet appearing on pages that shouldn't contain the switcher, as we are still merging refdocs from multiple generation tools.  See b/194329328 for more details.

Adding the switcher will be handled by the switcher script during the documentation staging process.

Note - the vast majority of changes in this CL are testData changes

Bug: 194329328
Bug: 193907212
Test: ./gradlew test
Change-Id: If9d62b2f6d460059d28b958c7af5996a7f549a9c