Autoupdate of 2 submodules
*system/tools/aidl updated to f2fdd1d54b826aed592fc3d162e44043f11ae65b
Rework unit tests for lazy AIDL services.

With this commit, we introduce a new service that are used for testing
the active services callback.
This fixes a potential race in the tests and also introduces a better
way to check if the callback has effectively been run within the
service.
We pass a file descriptor to the service that is then used to notify the
test that the callback has successfully run and the service is about to
exit.
Previous tests only relied on the observation from the unit test that
the service is not running anymore.

Bug: 191781736
Test: atest aidl_lazy_test
Change-Id: Id3d994bbb8634c2ae049c9c9e640673d83795f80

*system/tools/hidl updated to 3004e180e57a4b45a26bc6cf688769bf31243954
Rework unit tests for lazy HIDL services.

With this commit, we introduce a new service that are used for testing
the active services callback.
This fixes a potential race in the tests and also introduces a better
way to check if the callback has effectively been run within the
service.
We pass a file descriptor to the service that is then used to notify the
test that the callback has successfully run and the service is about to
exit.
Previous tests only relied on the observation from the unit test that
the service is not running anymore.

Bug: 191781736
Test: atest hidl_lazy_test
Change-Id: Ifa0761f0f16862afc6b154abd151886d53ab1489