abitidy: add pass to drop naming-typedef-id attributes
The attribute naming-typedef-id is a backwards link from an anonymous
type to the typedef that refers to it. This attribute only exists for
struct/class and not union or enum types. It is also ignored by
abidiff.

Unfortunately, libabigail sometimes conflates multiple anonymous types
that have naming typedefs and only one of the typedefs can "win". ABI
XML is thus sensitive to processing order and can also end up
containing definitions of an anonymous type with differing
naming-typedef-id attributes.

It's best to just drop the attribute.

	* tools/abitidy.cc (discard_naming_typedefs): New
	function to traverse the XML document and remove
	naming-typedef-id attributes from class-decl elements.
	(main): Add discard-naming-typedefs option handling and
	action.

Bug: 194371728
Signed-off-by: Giuliano Procida <gprocida@google.com>
Change-Id: I476085879c0fa3ce7c11c526c90726da9d88f348