Broaden the checks of the FuchsiaDevice during FlashTest
Broaden the exception in the Flash test to catch any exception
and if the testbed has PDUs, the reboot the device to prevent
it from being in a DeviceOffline state.  Moved the boot into
bootloader command to just before flashing so if download the
images fail, the device will not be stuck in fastboot.  This
is for fxb/81465

Bug: None
Test: Ran FlashTest
Change-Id: I9ed4d2eb6c949129783cc7b6f21e595b783d4109