Autoupdate of 2 submodules
*frameworks/base updated to 4eaa49eeb615d129ea6174f20d2373d9eba84452
Merge "Add a system API UpdateEngine to switch the slot back and forth"
*system/update_engine updated to da607a35bd2ea112844c7500aa21dc08127d9bf3
Add an API in update_engine to switch the slot back and forth

In some cases, the OTA update want to flip the slot to source
without cancelling the entire update. The functionality was
achieved by ResetStatus(). However ResetStatus() will clean up
the update progress and free up space to support virtual A/B.

Test: build
Bug: 187321613
Change-Id: I508c8034e69ad9f3159b985241451b0f5806bab7