vendor: true makes the interface to be not available for apps
There are several AIDL interfaces that has `vendor: true`. They shound't
be built with `sdk_version`, otherwise the vendor variant won't be
created [1]. To fix the problem, an AIDL interface with `vendor: true`
is not available for apps so that sdk_version is not set for the
interface. This makes sense because 'vendor: true' means the library is
exclusively for vendors and therefore can't be available to apps by any
means.

[1]
https://cs.android.com/android/platform/superproject/+/master:build/soong/cc/image.go;drc=48f3778cb4b86a3d5827ba4b020157ace975917b;l=515

Bug: 161456198
Test: m
Change-Id: I0e3d3aa105d2280458df11dda9b7d61c8bcf2c8f