Update CaptivePortalLoginTests TEST_MAPPING
Move CaptivePortalLoginTests from postsubmit to presubmit, as recent
runs have been green.

Also fix the mainline-postsubmit configuration: the build rule of the
test needs to specify test_mainline_modules to match the TEST_MAPPING
configuration, and the test needs to be in general-tests. This test
does not need to be in device-tests, as it does not use a
device-specific signing key.

Bug: 162296431
Test: TH needs to test test mapping change

Change-Id: Idb89dd318958697e2e58e7d6ca3a7e8633550607