db845c.mk: Do not copy kernel modules to vendor image
Copy the kernel modules the ramdisk/vendor_ramdisk only.

Right now we keep a copy of all the kernel modules in
/vendor/lib/modules, which results in modprobing of the
modules from /vendor/lib/modules as well and though they
fail safely (well most of the time and they should) but we
have run into a regression where dual loading of Qcom UFS
module resulted in probing failures. Ideally dual loading
should not reach up to probing stage and there might be
something more to that crash, but since this issue is
difficult to reproduce and debug, the easiest way out is
to make sure we do not try to load a kernel module twice.

Signed-off-by: Amit Pundir <amit.pundir@linaro.org>
Change-Id: I1be3bccd1ababaf258f1f87c75636919b962ad8d