Create a tool to help share common ELFs.
Created share_common_elfs.sh. This bash script is intented to be used on
ELF(s) in the offline dump directories created by unwind_for_offline
with the goal of eliminiating copies of the same ELF files accross the
subdirectories of 'libunwindstack/offline_files/'.

See the usage function in the tool for additional details.

Bug: 192012600
Test: Manual testing:
* Invalid usage (e.g. share_common_elfs -a /path/to/specific/elf) for
each option flag.
* All valid combination of cmd line flags. Where the 'common/' dir
either did not contain the specified elf or already contained it.
* Path to invalid and non-ELF files.

Change-Id: I31a65be455628c52d63d07aaa0218200531c2d2d