default to media ready but show all supported sizes
If the user wants to select a media size that their
printer is not reporting over media ready, we should
allow them to do so if that media size if supported.
This change makes the first found media ready size the
pre selected default media size in the print flow, but
allows the user to select any supported media size if
they want to.

Test: Print to a media ready printer and check default media size
Signed-off-by: Kurtis Schlepp <mopriadevteam@gmail.com>
Change-Id: I946e408b9d2c63c674ced6cd34278f6fa4e7d6ba