floss: Start Media (a2dp + avrcp) on btadapterd init
We want to start the core stack, and profile services, when we start the
stack; we don't want to expose those to clients.

Note that during init, events happens synchronously, so we can avoid the
race condition that the client wants to init A2dp while core stack is
not ready.

Tag: #floss
Bug: 195398850
Test: run btadapterd
Change-Id: I4834071b14dac86c4d25aacfc3faa2d174457d9e