vts: kernel: toolchain: relax LLD requirement to S+
Do to a major mistake of mine, even though R+ had all of the kernel
patches necessary to link the kernel with LLD, I did not cherry-pick the
following patches to the necessary rvc-dev branch after landing them in
master. Device owners now upgrading devices launched with R are finding
that they're failing this test when upgrading to S.
- aosp/1281111
- aosp/1288903

Relax the linker test, but don't relax the compiler tests; we don't have
reported failures of that yet. I can relax those too, if needed, but I'd
prefer to wait and see if anyone is affected (non-aarch64).

Bug: 153484457
Bug: 153484457
Bug: 154354791
Bug: 194366277
Test: mm && adb sync data && \
      adb shell /data/nativetest/vts_kernel_toolchain/vts_kernel_toolchain
Signed-off-by: Nick Desaulniers <ndesaulniers@google.com>
Change-Id: I1b47b58a6d29258affa8ae9c865634665321fdce
Merged-In: I1b47b58a6d29258affa8ae9c865634665321fdce