Add a make_injection repository rule.
The repository rule enables soong_ui --make-mode to run before Bazel
automatically, as a repository rule. This allows artifacts like host
tools to be built and used in the Bazel build graph without having to
convert them in a bottom-up manner.

For a start, this is used for exposing the APEX host tools in the apex
rule.

Test: bp2build.sh (apexes are built in //system/timezone/apex/... and //build/bazel/examples/minimal/apex/...)
Change-Id: I4392b92f2c37c3d04c282dd6919789756da98ad6