Single-thread consecutive offline benchmark/tests.
Create the diverse PCs of the same thread process profiler-like
(consecutive) offline unwind benchmarks.

The offline unwind dumps/samples gathered/selected for this benchmark are
a diverse set of PCs from the thread that lies within libbluetooth.so at
the time of the unwind to attempt to mimic a high-level of repition that
a profiler like Perfetto or Simpleperf may encounter when profiling a
single process.

Benchmarking offline unwinds are critical for performance analysis
regarding potential bottlenecks that affect clients includeing Perfetto
and Simpleperf.

See b/192012600 for additional information regarding these benchmarks.

Bug: 192012600
Test: Benchmarking/test CL. Existing unit tests still pass.
Change-Id: Idfde9781e28be698a42db04135a115b5d71d3232