Skip ApexShimUpdateTest if shim APEX is missing.
This should never happen, but the fact that test became flaky suggests
that it does happen ¯\_(ツ)_/¯.

I suspect that this is related to the fact that we use `adb remount` to
push test APEXes on a device. Once I'm done with refactoring
apexservice_test we should be able to remove the dependency on `adb
remount`, but for the time being we can just skip the tests instead of
failing.

Bug: 196369686
Test: atest ApexTestCases
Change-Id: I67e41c419058c1d57a13da82c26e5e04a2bd4a57