idsig: less copies when creating the merkle tree
This change improves the merkle tree generation routine so that we don't
do unnecessary data copies. Previously, hashes for level N is written
to a temporary store and then copied into the tree. Even worse, the
hashes written to the tree is copied into another buffer when they are
used as the inputs for the next level.

With this CL, the hashes are directly written to and read from the tree.
This is done by having two (non-overlapping) slices on the hash tree.

Bug: N/A
Test: cargo test
Change-Id: I34be81ece6941eba78980c8bc4697ed5d523ed53