Do not acquire runtime_shutdown_lock_ in Abort()
Abort can be called, particularly in OOM situations, when we already
hold the lock. Abort() should minimize the locks it acquires.

This is intended to be a minimal, low-risk change. Generated code
should be essentially unchanged, except in Abort(). This does not
address the question of whether IsShuttingDown really needs to
lock at all.

Test: Build and boot AOSP.
Bug: 195884830
Change-Id: I0ee4a7ca7348153436fec0fecc1d1f2ca1f7a30c