Fix a conflict between the bc source and Apple's fortify implementation.
Reported upstream, and will remove this when we agree on the real fix.

(Note that we already don't have fortify for glibc or musl, so this
isn't as bad a step backwards as it seems, and at least this one should
be easily and quickly rectified. If this makes you sad, it's Linux you
should be worried about...)

Bug: https://github.com/gavinhoward/bc/issues/35
Test: postsubmit :-(
Change-Id: I73b7b4bf874ec5b7cea9f8a07a8c5ad4edb2f71e