Add unit tests for ProcessTracer.
Created unit tests for the ProcessTracer utility class. As mentioned in
the stop in specific ELF CL
(https://android-review.googlesource.com/c/platform/system/unwinding/+/1769946)
ProcessTracer could be used by existing libunwindstack unit tests as
well as future libunwindstack tools and tests.

Due to the limited time left in the internship, this CL focuses on
including common success and failure cases for ProcessTracer. The primary
benefit from this CL is the test fixture that provides the necessary
"infrastructure" to create additional ProcessTracer unit tests in the
future.

In order to test stopping in a desired ELF, two other refactors were
made:
1. Moving the code to dlopen the local ELF from
   LocalUnwinderTest::unwind_after_dlopen into TestUtils.* so this test
   can re-use the code to dlopen libunwindstack_local.so.
2. Adding a busy wait loop into TestLocal.cpp so ProcesTracer finds one
   of the child process threads within libunwindstack_local.so.

Bug: 196383955
Test: ProcessTracerTest.cpp (this CL).
Change-Id: I6a177a6de19fc1f736dbda1faf46c939722710d1