Allow configuring acloud URL
The device action "Local Virtual Device Setup" defines acloud as a test resource. Its option "--gce-driver-path" refers to the resource in the working directory. The URL defaults to the prebuilt acloud in the container file system. The user can change the URL if the prebuilt acloud has compatibility problems with the latest virtual devices.

PiperOrigin-RevId: 390748501
Change-Id: I8c7fbdb67bc5ff65bcbefc7c202f4a91f7461305