Autoupdate of 2 submodules
*frameworks/libs/net updated to 74edf5abe40649c730c4cd9d2913cda4e92f56ad
Add the ConnectivityModuleTest annotation

The annotation will be used to mark tests cases that are part of MTS
modules used by multiple modules, to indicate that the test case should
only be run if the MTS suite should cover an updated Connectivity
(tethering) module.

This allows tests to verify new behavior in the Connectivity module,
without failing if that module is not updated.

Bug: 196755836
Test: mts-tradefed run mts-network with exclude-annotation in the config
Change-Id: I05d2200c7ab68042e4e48029e045f164c26b8037

*test/mts updated to f5a1e49fa7893d017da57b29723194f1890f8e61
Exclude ConnectivityModuleTest in NetworkStack MTS

Tests annotated with ConnectivityModuleTest should only be run in the
connectivity (tethering) MTS, not in NetworkStack/CaptivePortalLogin
MTS.

Bug: 196755836
Test: mts-tradefed run mts-network
Change-Id: Ib130e94b87a4e12ddd59aae7c6437e24d744ecad