Remove the confusing log in getPrefix64
This log can sometimes confuse people. A common case where this
log is printed is that an app specifies AF_UNSPEC but there are
no resolved IPv6 addresses for the hostname.

ResolverController: No valid NAT64 prefix (106, <unspecified>/0)

After this change, the resolver logs will still be enough for us
to infer if DNS64 synthesis is performed.

Bug: 196888805
Test: build dnsresolver pass
Change-Id: Ic06299d1a0c6526e7ed1a0a418a8da19dbf2b6da