suspend_blocker: Use Cuttlefish's workaround to avoid suspending the device
Reuse Cuttlefish's suspend blocker service to avoid
suspending the device while running CTS tests.

This is actuallly a cherry-pick of Amit's change on db845c here:
https://android-review.googlesource.com/c/device/linaro/dragonboard/+/1723130

The problem is that "adb shell stop" is causing a PM suspend and
subsequent loss of adb connectivity. The problem could also be reproduced by
the vts ltp test.

Test: booted to homescreen with hikey and hikey960 board,
      and adb shell stop does not cause suspend any more

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I290faa5182135a619c4053d12a88a2d534641d7f