Remove direct uses of C++ compiler from run-test build
Avoid calling the system compiler as part of run-tests.
These uses are making porting to soong build tricky.

It would be possible to build the generators in soong,
but it is much simpler to just rewrite them in python.
This is in line with what majority of other run-tests do.

Bug: 147814778
Test: test.py --host
Change-Id: I97cd54c85db81d70795bd5321a452c3306f06b72