Fix odsign_e2e_tests to accurately get file modified time in ms.
Before this change, we were using "stat -c '%.3Y'" to get file modified
time in ms. However, the Toybox's `stat` implementation truncates to
timestamp to seconds, making the timestamp not accurate enough, and
causes the test to be flaky. Flakes observed when odrefresh compiles
artifacts right after a boot in a test, where the compilation time and
the boot time only differ in milliseconds.

This change fixes the problem by using "stat -c '%y'" and converting the
formatted time into timestamp in Java.

Bug: 193616266
Test: atest odsign_e2e_tests
Change-Id: I36c571e9cafc2b4e99ae20f196a138834d7f57e0