Do not use ScopedResource for FindFirstFile handle
The current usage of ScopedResource does not work for Android Studio
Profiler, because INVALID_HANDLE_VALUE under their compile configuration
is defined as ((HANDLE)(LONG_PTR)-1), which is a reinterpret_cast, and
reinterpret_casts cannot be used in constexpr. Template arguments are
required to be constexprs.

Since there is just one use case of this, it's easier to call FindClose
directly. If we have many more uses of FindFirstFile handle in the
future, we can look into defining a different scoped object for it.

Bug: 182165266
Change-Id: I27bc917643b1b13d4bfd4c1ad9299ee3324a2b08