Autoupdate of 4 submodules
*external/rust/crates/ring updated to 199b0c1509138849cd41191db34233347deb84fa
Make ring available to the virt APEX

Bug: 193504400
Test: m
Change-Id: I9d1abbef447d4f511e24f3cf4b57d9894a6adcc3

*external/rust/crates/spin updated to 9ea9c8c329f990dade80d4f83e55511b64ca60d7
Make spin available to the virt APEX

Bug: 193504400
Test: m
Change-Id: Ieee8a41591179b51f5ed2bf4a1a29de9648a63ea

*external/rust/crates/untrusted updated to d0de1ba30b66c3155f4b13cb11a82abb16525827
Make the untrusted crate available to the virt APEX

Bug: 193504400
Test: m
Change-Id: I723c19514bdfda6cf92fe3ea57eb70e4b788a69b

*packages/modules/Virtualization updated to 0a2484304d22c9350ffc0ae805c7794c9302769a
Create idsig file automatically

Before a VM is started, the idsig file is created (or updated) by the
virtualization service. This is needed because the idsig file is usually
not available, especially when the APK is downloaded from the store.

Note that the generated idsig file is not a signed one. Therefore, the
APK is first verified using the APK signature scheme V3 (or V2) over a
dm-verity device backed by the APK and the merkle tree (and root hash)
from the idsig file. Only if the verification is successful, the root
hash stored to the instance.img and then used for the subsequent boots
of the VM.

Bug: 193504400
Test: atest MicrodroidHostTestCases
Test: run MicrodroidDemoApp without having the idsig file in
/data/local/tmp/virt.

Change-Id: I9fad05ca9562ae0666431102a8147d0f76f04e6a