Remove ProcessLifecycleOwnerInitializer workaround from manifest
The androidx.lifecycle.ProcessLifecycleOwnerInitializer provider
was added to the manifest as a workaround for incorrect manifest
merging in b/112587252.  Soong has been fixed to only merge the
manifests at the final APK, preserving the right value of
${applicationId}, so the workaround is no longer necessary.

An upcoming drop of androidx to the platform removes the
androidx.lifecycle.ProcessLifecycleOwnerInitializer in favor
of androidx.startup.InitializationProvider, which causes failures
at runtime if the workaround is not removed.

Bug: 197536008
Test: boot with new androidx drop
Test: ProcessLifecycleOwnerInitializer still in compiled AndroidManifest.xml
Change-Id: I49588c6d5627b66ded2219dc358b52568822ee38