ART: Change error-reporting of run-test-jar
In the run-test-jar script used by the host and target test scripts,
certain error-exit messages are displayed. When running the host and
target test scripts, these error-exit messages are easy to miss.

The error-exit messages are now passed to a new function called
"error_msg()" to be displayed in bold and red to make them easier
to spot. Additionally, the exit statement is called within the function
to avoid code duplication.

Test: ./art/test/run-test --gdb --gdbserver --host 491-current-method

Change-Id: I6929aaa971f48e04918152042596f3aaa988c609