Enable window resizing with guest Vulkan.
The problem with this is that after resizing, if Android does not render
a new frame, the emulator would keep showing the black screen. This is
unlikely an issue for gameplay since games are always rendering new
frames.

Instead of recreating the swapchain, we could make viewport a dynamic
state instead, which requires re-recording command buffers. That
approach doesn't address the black screen issue either. Hence, we prefer
the simpler approach for now.

Test: Run the emulator with VVL enabled, and test resizing the window.
Bug: 189866933
Change-Id: I5aff9cc017521fdb20bb5e6328137fd56430d3db