Autoupdate of 2 submodules
*build/soong updated to 7e4c05fcc8df19d7930d2dc4bb35f4da971cc625
Merge "Add exclude_runtime_libs to more targets."
*system/unwinding updated to efe72cbfbd9bd1d118780994b5cf364054a4b98a
Register a runtime dependency on libdexfile in all variants that use
libdexfile_support.

Since Soong doesn't propagate transitive dependencies for static
libraries, it is necessary to add this dependency explicitly for all
variants that build with the static libdexfile_support.

This required extending the support for exclude_runtime_libs in Soong.

Test: deapexer info \
      out/target/product/vsoc_x86_64/system/apex/com.android.runtime.apex
  Check that it lists libdexfile.so as a dependency.
Test: adb root && adb shell 'kill -SEGV `pidof zygote64`'
  Check in logcat that the crash_dump(32|64) process doesn't print
  "dlopen failed: library "libdexfile.so" not found"
Bug: 197856821
Change-Id: I1b9dafb8818e026cdd757888861fdf6208c9e465