Implement runTask to schedule tasks in PostWorker
Use std::packaged_task<void()> to wrap the task, and use m_runOnUiThread
to schedule the task on UI thread if m_mainThreadPostingOnly is true.

With this improvement, the caller now doesn't have to manually encode
the parameters and send parameters via m_toUiThread. Now the caller can
directly capture the parameter in a lambda.

Test: compile and run the emulator
Change-Id: Ic4304643377287891aa3014aea7c9fac8be409f6