vts: ltp: Disable controllers.memcg_control_* tests from vts runs
Since CONFIG_MEMCG was enabled in gki_defconfig, we've been seeing
increased failures that we've tracked down to a side effect from
problems with the memcg_control test.

You can find more details here:
  https://issuetracker.google.com/issues/197942864

But the gist is due to issues in both toybox and the test
itself, the test often will leave processes running resulting
in the control script going into zombie state. After 10 minutes
llkd will jump in and kill adbd which causes a number of the
following tests to fail.

So this patch disables the test from vts for now until fixes
land upstream and are merged back into AOSP

Bug: 197942864
Signed-off-by: John Stultz <john.stultz@linaro.org>
Change-Id: If30b11c85136c0ea4fa982e848573d9b82bb5707