Allow the check command to use globs.
An earlier commit allowed the update command to take a glob, making it
easier to update multiple repositories in a single invocation.  This
moves that code into a shared function and lets the check command use
it as well, giving it the same benefits.

It also slightly changes the ordering: instead of sorting all paths,
it only sorts the results of the glob (which are not deterministic).

Finally, since I was refactoring code, I also refactored the
report-generation code to be shared across the two commands.

Test: Check and upgrade multiple repositories using a * wildcard.
Change-Id: I5b53d5c258d52e8dba9f7afa6f3f169808c86be8